Fixes related to deleting records

+++ js/dev-ctrl.js
@@ -400,6 +400,22 @@ DevCtrl.DataService.factory = ['$window', '$http', '$mdToast', '$timeout', 'sock
             deleteRow : function(row) {
                 var resource = "data.php/" + row.tableName + "/" + row.id;
 
+                // Check for foreign key constraints
+                var referencedTable = false;
+                angular.forEach(row.referenced, function(refTable, refs) {
+                    if (Object.keys(refs).length > 0) {
+                        //TODO: cannot delete value due to foreign key constraint
+                        referencedTable = refTable;
+                    }
+                });
+
+                if (referencedTable) {
+                    var msg = "Cannot delete " + row.tableName + " record due to foreign key constraint on " + referencedTable;
+
+                    self.errorToast({error: msg});
+                    return;
+                }
+
                 $http.delete(resource)
                     .success(function (data) {
                         self.loadData(data);
@@ -679,14 +695,16 @@ DevCtrl.DataService.factory = ['$window', '$http', '$mdToast', '$timeout', 'sock
                         var pk = self.getSchema(table).pk;
 
                         angular.forEach(tableData, function(value, key) {
-                            if (angular.isString(pk)) {
-                                delete dataModel[table].indexed[key];
-                            }
-                            else {
-                                angular.forEach(value, function(val2, key2) {
-                                    delete dataModel[table].indexed[key][key2];
-                                });
-                            }
+                            // Remove references
+                            var record = dataModel[table].indexed[key];
+
+                            angular.forEach(record.foreign, function(referenced, refID) {
+                                if (angular.isDefined(referenced.referenced[table][key])) {
+                                    delete referenced.referenced[table][key];
+                                }
+                            });
+
+                            delete dataModel[table].indexed[key];
                         });
 
                         // Rebuild object list
@@ -1510,9 +1528,6 @@ DevCtrl.Panel.Directive  = ['$mdDialog', 'DataService', function($mdDialog, Data
         controller: function($mdDialog, DataService) {
             var self = this;
             this.fields = this.panelObj.fields;
-            this.pcontrols = this.panelObj.referenced.panel_controls;
-
-            this.appConfig = DataService.config;
 
             this.addControl = function($event) {
                 $mdDialog.show({
@@ -1534,7 +1549,7 @@ DevCtrl.Panel.Directive  = ['$mdDialog', 'DataService', function($mdDialog, Data
             };
 
             this.setAllSwitches = function(val) {
-                angular.forEach(this.pcontrols, function(pcontrol) {
+                angular.forEach(self.panelObj.referenced.panel_controls, function(pcontrol) {
                     var control = pcontrol.foreign.controls;
 
                     if (control.fields.usertype == 'switch') {
